+++ tests/integration/hook-order2/test.php
@@ -10,8 +10,8 @@ Hooks::register( 'Something', 'HookOrder2\SecondClass::hook2' );
 $tainted = $_GET['user'];
 $output = '';
 Hooks::run( 'Something', [ $tainted, &$output ] );
-echo $output;// XSS
-htmlspecialchars( $output ); // DoubleEscaped
+echo $output;// XSS caused by 10, 12, 21 (in this order)
+htmlspecialchars( $output ); // DoubleEscaped caused by 12, 18 (in this order)
 
 class SecondClass {
 	public static function hook1( $arg1, &$arg2 ) {