+++ tests/integration/passbyrefnooverride/test.php
@@ -39,7 +39,7 @@ $extraArg2 = 'Foo';
 
 $var5 = $_GET['unsafe'];
 unsafe( $var5 );
-echo $var5; // Unsafe TODO: should NOT have line 40 in the caused-by
+echo $var5; // Unsafe, must NOT have line 40 in the caused-by
 
 function maybeReplaceTaint( &$arg ) {
 	if ( rand() ) {