+++ tests/integration/passbyrefnooverride3/test.php
@@ -6,7 +6,7 @@ function identicalSides( &$arg ) {
 
 $var1 = $_GET['baz'];
 identicalSides( $var1 );
-echo $var1; // Unsafe, caused by lines 8, 7 and 4
+echo $var1; // Unsafe, TODO caused by lines 7, 8 and 4 (in this order)
 
 
 function indirect( &$arg ) {
@@ -16,7 +16,7 @@ function indirect( &$arg ) {
 
 $var2 = $_GET['baz'];
 indirect( $var2 );
-echo $var2;// Unsafe, caused by lines 17, 18, 14, 13 (TODO: in this order)
+echo $var2;// Unsafe, TODO caused by lines 17, 18, 14, 13 (in this order)
 
 
 function usingList( &$arg ) {
@@ -26,7 +26,7 @@ function usingList( &$arg ) {
 
 $var3 = $_GET['baz'];
 usingList( $var3 );
-echo $var3;// Unsafe, caused by lines 27, 28, 24, 23 (TODO: in this order)
+echo $var3;// Unsafe, TODO caused by lines 27, 28, 24, 23 (in this order)
 
 
 function ternary( &$arg ) {
@@ -36,7 +36,7 @@ function ternary( &$arg ) {
 
 $var4 = $_GET['baz'];
 ternary( $var4 );
-echo $var4;// Unsafe, caused by lines 37, 38, 34, 33 (TODO: in this order)
+echo $var4;// Unsafe, TODO caused by lines 37, 38, 34, 33 (in this order)
 
 
 function indirectLink( &$arg ) {